adjusted App.js for DigitButtons and OperationButtons.
Added DigitButtons.js (handles logic and structure):
This component renders a calculator button for a single digit.
When clicked, it dispatches an ADD_DIGIT action to the reducer in App.js.
The digit passed in (0–9 or .) is added to the current operand.

it’s the reusable building block for all number buttons in your calculator.

Added OperationButton.js(handles logic and structure):
This component renders a calculator button for an operation (+, -, *, ÷).
When clicked, it dispatches a CHOOSE_OPERATION action to the reducer in App.js.
The chosen operation is stored in state so the calculator knows which arithmetic function to apply.
it’s the reusable building block for all operator buttons.

Added styles.css(handles styling):
- built alongside the other components to adjust styling, position
- it defines the layout (grid system), colors, fonts, spacing, and hover effects...

Added index.js:
-boots the app